move pki_utils.h to private/ (part2)
this should never have been public
and now we can simply include windows.h
instead of re-typedef'ing everything
